Change naming mechanism

+++ main.py
@@ -12,7 +12,7 @@ class Game:
     db.saved_games_database()
 
     def __init__(self, name):
-
+        if name == '' : return exception
         name = ('game_' + db.remove_spacing(name)).lower()
         self.designated_db_table = f"{name}"           ##And initialize database
         db.create_game_database(name)
@@ -66,9 +66,13 @@ def main():
 
     if player_choice == 'new':
 
-        while True :
-            try:
-                game = Game(input("Enter your game name: "))               ##Rewrote this part using the saved_games list
+        while True:
+            name = input("Enter your game name: ")
+            if name == '' :
+                print("The name cannot be empty\n")
+                continue
+            try :
+                game = Game(name)
                 break
             except :
                 print("Please only enter characters from a..z and numbers 0..9")
